requirements.txt
PyPDF2
reportlab
nltk
google-generativeai
pinecone
